refactor: create exorcism module, move relevant items to it
