+++ src/bundles.rs
@@ -9,8 +9,3 @@ pub struct GrassBundle {
     pub sprite_sheet_bundle: SpriteSheetBundle,
     pub wind_timer: WindTimer,
 }
-
-#[derive(Clone, Bundle, LdtkIntCell)]
-pub struct ExorcismBlockBundle {
-    pub exorcism_block: ExorcismBlock,
-}
