+++ src/gameplay/mod.rs
@@ -7,8 +7,3 @@ pub mod transitions;
 pub fn xy_translation(coords: IVec2) -> Vec2 {
     Vec2::new(coords.x as f32 + 0.5, coords.y as f32 + 0.5) * UNIT_LENGTH
 }
-
-#[derive(Copy, Clone, Eq, PartialEq, Debug, Hash)]
-pub struct DeathEvent {
-    pub willo_entity: Entity,
-}
